Moved apply confirmation to apply folder
