+++ apps/site/src/app/apply/sections/ApplyConfirmation/ConfirmationDetails.module.scss
